✅ 修复 winconpty smoke test 在非 Windows 平台的编译错误
TestConPtyReadsChildOutput 引用了仅在 conpty_windows.go (//go:build windows)
定义的 Start / IsConPtyAvailable / ConPty 等符号,但它原先放在无构建约束的
creation_flags_test.go 中,导致非 Windows 平台(CI ubuntu-latest 上的
go test ./... 与 golangci-lint)编译失败。

把该 smoke test 拆到新的 conpty_windows_test.go (//go:build windows),
creation_flags_test.go 仅保留跨平台的 flag 断言测试。
